Module zkpass_core::interface

Structs

DataVerificationRequest

The DataVerificationRequest captures all information contained in a DVR.

KeysetEndpoint

The KeysetEndpoint is used to represent the JWKS (JSON Web Key Set) end points.

PublicKey

The PublicKey is used to represent a public key.

VerifiedNestedTokenData

VerifiedNestedTokenDvr

ZkPassProof

The ZkPassProof is the value returned by the generate_zkpass_proof API of the zkPass Service.

Enums

PublicKeyOption

The PublicKeyOption variant provides alternative ways to represent a public key.

ZkPassError

The ZkPassError enum defines the various errors that come from the zkPass Service.

Traits

KeysetEndpointResolver

The KeysetEndpointResolver trait defines a function for retrieving a public key from a JWKS endpoint.

Functions

decrypt_jwe_token

encrypt_data_to_jwe_token

get_current_timestamp

Gets the number of seconds that has elapsed since the Unix Epoch time.

get_keyset_endpoint_params

sign_data_to_jws_token

tokenize_data

verify_data_nested_token

verify_dvr_nested_token

verify_jws_token